moved config to gpio package and made it private
